Improved user interface in index.html and pbd.html
Moving the "Next" and "Previous" buttons: The buttons have been removed from the browser and placed directly below the book display for improved accessibility.
Added sharing popup: When the "Share" button is clicked, a popup is displayed to enable sharing on social networks.
Upgrade to Bootstrap 5.3.3: Use the latest version of Bootstrap to benefit from the latest features and fixes.